src/requirements.txt
aiogram==2.23.1
aiohttp==3.8.3
aiosignal==1.3.1
aiosqlite==0.17.0
anyio==3.6.2
APScheduler==3.9.1.post1
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.1.0
Babel==2.9.1
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
Deprecated==1.2.13
frozenlist==1.3.3
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.4
iso8601==1.1.0
magic-filter==1.0.9
multidict==6.0.2
packaging==21.3
pydantic==1.10.2
pyparsing==3.0.9
pypika-tortoise==0.1.6
python-telegram-bot==20.0a4
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
redis==4.4.0rc4
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
tornado==6.2
tortoise-orm==0.19.2
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
wrapt==1.14.1
yarl==1.8.1
